+++ data/macros/warlock.lua
@@ -1,0 +1,6 @@
+local Deathless = Deathless
+
+Deathless.Data = Deathless.Data or {}
+Deathless.Data.Macros = Deathless.Data.Macros or {}
+
+Deathless.Data.Macros.Warlock = {}